Revised data types

+++ src/ft_reader.h
@@ -1,22 +1,23 @@
 #ifndef FT_READER_H
 #define FT_READER_H
 
+#include <stddef.h>
 #include <stdbool.h>
 
 struct ft_reader {
     unsigned char *buffer;
-    int bufferLength;
-    int offset;
-    int bytesRead;
-    int messageLength;
-    int offsetChunk;
+    size_t bufferLength;
+    size_t offset;
+    size_t bytesRead;
+    unsigned int messageLength;
+    size_t offsetChunk;
 };
 
 struct ft_reader *ft_readerCreate();
 // Returns -1 if chunk couldn't be processed
-int ft_readerRead(struct ft_reader *reader, unsigned char *chunk, int effectiveChunkLength, unsigned char **buffers);
-bool ft_readerReadMessageLength(struct ft_reader *reader, unsigned char *chunk, int effectiveChunkLength);
-bool ft_readerReadMessageContent(struct ft_reader *reader, unsigned char *chunk, int effectiveChunkLength);
+int ft_readerRead(struct ft_reader *reader, unsigned char *chunk, size_t effectiveChunkLength, unsigned char **buffers);
+bool ft_readerReadMessageLength(struct ft_reader *reader, unsigned char *chunk, size_t effectiveChunkLength);
+bool ft_readerReadMessageContent(struct ft_reader *reader, unsigned char *chunk, size_t effectiveChunkLength);
 void ft_readerCreateBuffer(struct ft_reader *reader);
 void ft_readerDestroy(struct ft_reader *reader);
